// dashboard/src/forms/types/Fieldnames.ts
export enum FieldName {
  ACME_CONFIG = "acmeConfig",
  ACME_CONFIG_AGREETOS = "acmeConfig.agreeTos",
  ACME_CONFIG_APPROVE_DOMAIN = "acmeConfig.approveDomain",
  ACME_CONFIG_COMMUNITY_MEMBER = "acmeConfig.communityMember",
  ACME_CONFIG_CONFIG_DIR = "acmeConfig.configDir",
  ACME_CONFIG_DEBUG = "acmeConfig.debug",
  ACME_CONFIG_EMAIL = "acmeConfig.email",
  ACME_CONFIG_SECURITY_UPDATES = "acmeConfig.securityUpdates",
  ACME_CONFIG_SERVERNAME = "acmeConfig.servername",
  ACME_CONFIG_SERVER = "acmeConfig.server",
  ACME_CONFIG_TELEMETRY = "acmeConfig.telemetry",
  ACME_CONFIG_VERSION = "acmeConfig.version",
  ARGS_TRANSPORT = "argsTransport",
  ARGS_TRANSPORT_COLORIZE = "argsTransport.colorize",
  ARGS_TRANSPORT_HANDLE_EXCEPTION = "argsTransport.handleExceptions",
  ARGS_TRANSPORT_JSON = "argsTransport.json",
  ARGS_TRANSPORT_LEVEL = "argsTransport.level",
  ARGS_TRANSPORT_TIMESTAMP = "argsTransport.timestamp",
  AUTH_TIMESTAMP_CACHE_SIZE = "authTimestampCacheSize",
  AWS_CREDENTIALS = "awsCredentials",
  AWS_CREDENTIALS_ACCESS_KEY_ID = "awsCredentials.accessKeyId",
  AWS_CREDENTIALS_ENDPOINT = "awsCredentials.endpoint",
  AWS_CREDENTIALS_SECRET_ACCESS_KEY = "awsCredentials.secretAccessKey",
  AWS_CREDENTIALS_SESSION_TOKEN = "awsCredentials.sessionToken",
  AZ_CREDENTIALS = "azCredentials",
  AZ_CREDENTIALS_ACCOUNT_KEY = "azCredentials.accountKey",
  AZ_CREDENTIALS_ACCOUNT_NAME = "azCredentials.accountName",
  BUCKET = "bucket",
  CACHE_CONTROL = "cacheControl",
  DISK_SETTINGS = "diskSettings",
  DISK_SETTINGS_STORAGE_ROOT_DIRECTORY = "diskSettings.storageRootDirectory",
  DRIVER = "driver",
  ENABLE_HTTPS = "enableHttps",
  GC_CREDENTIALS = "gcCredentials",
  GC_CREDENTIALS_CREDENTIALS = "gcCredentials.credentials",
  GC_CREDENTIALS_CREDENTIALS_CLIENT_EMAIL = "gcCredentials.credentials.client_email",
  GC_CREDENTIALS_CREDENTIALS_PRIVATE_KEY = "gcCredentials.credentials.private_key",
  GC_CREDENTIALS_EMAIL = "gcCredentials.email",
  GC_CREDENTIALS_KEY_FILENAME = "gcCredentials.keyFilename",
  GC_CREDENTIALS_PROJECT_ID = "gcCredentials.projectId",
  HTTPS_PORT = "httpsPort",
  MAX_FILE_UPLOAD_SIZE = "maxFileUploadSize",
  PAGE_SIZE = "pageSize",
  PORT = "port",
  PROOFS_CONFIG = "proofsConfig",
  PROOFS_CONFIG_PROOFS_REQUIRED = "proofsConfig.proofsRequired",
  READ_URL = "readUrl",
  REQUIRE_CORRECT_HUB_URL = "requireCorrectHubUrl",
  SERVERNAME = "serverName",
  TLS_CERT_CONFIG = "tlsCertConfig",
  TLS_CERT_CONFIG_CERTFILE = "tlsCertConfig.certFile",
  TLS_CERT_CONFIG_KEYFILE = "tlsCertConfig.keyFile",
  TLS_CERT_CONFIG_KEYPASSPHRASE = "tlsCertConfig.keyPassphrase",
  TLS_CERT_CONFIG_PFXFILE = "tlsCertConfig.pfxFile",
  TLS_CERT_CONFIG_PFXPASSPHRASE = "tlsCertConfig.pfxPassphrase",
  VALID_HUB_URLS = "validHubUrls",
  VALID_HUB_URLS_ITEMS = "validHubUrls.items",
  WHITELIST = "whitelist",
  WHITELIST_ITEMS = "whitelist.items",
}